Added more action for nested tag inspect form
- Hide delete tag switch / button when the current tag is root tag

+++ src/Endermanbugzjfc/NBTInspect/uis/forms/NumbericValueEditForm.php
@@ -36,7 +36,7 @@ class NumbericValueEditForm extends ValueEditForm {
 		$s = $this->getUIInstance()->getSession();
 		$t = $this->getUIInstance()->getSession()->getCurrentTag();
 		$f->addLabel(TF::YELLOW . 'Acceptable value range: ' . TF::BOLD . TF::GOLD . (Utils::getNumbericTagAcceptableValueRange($t) ?? TF::RED . 'UNKNOWN'));
-		$f->addSwitch(TF::RED . 'Delete tag');
+		if ($s->getRootTag() !== $s->getCurrentTag()) $f->addSwitch(TF::RED . 'Delete tag');
 		return $f;
 	}
 
@@ -50,12 +50,14 @@ class NumbericValueEditForm extends ValueEditForm {
 		}
 		array_shift($data);
 		array_shift($data);
-		if ($data[0]) {
-			$s->deleteCurrentTag();
-			$s->inspectCurrentTag();
-			return;
+		if ($s->getRootTag() !== $s->getCurrentTag()) {
+			if ($data[0]) {
+				$s->deleteCurrentTag();
+				$s->inspectCurrentTag();
+				return;
+			}
+			array_shift($data);
 		}
-		array_shift($data);
 		$t = $s->getCurrentTag();
 		switch (true) {
 			case $t instanceof FloatTag:
