+++ src/Endermanbugzjfc/NBTInspect/uis/forms/StringValueEditForm.php
@@ -32,7 +32,7 @@ class StringValueEditForm extends ValueEditForm {
 	protected function form() : \jojoe77777\FormAPI\Form {
 		$f = parent::form();
 		$s = $this->getUIInstance()->getSession();
-		$f->addSwitch(TF::RED . 'Delete tag');
+		if ($s->getRootTag() !== $s->getCurrentTag()) $f->addSwitch(TF::RED . 'Delete tag');
 		$t = str_split($s->getCurrentTag()->getValue(), 75);
 		foreach ($t as $v) $f->addInput('', '', $v);
 		return $f;
@@ -47,12 +47,14 @@ class StringValueEditForm extends ValueEditForm {
 			return;
 		}
 		array_shift($data);
-		if ($data[0]) {
-			$s->deleteCurrentTag();
-			$s->inspectCurrentTag();
-			return;
+		if ($s->getRootTag() !== $s->getCurrentTag()) {
+			if ($data[0]) {
+				$s->deleteCurrentTag();
+				$s->inspectCurrentTag();
+				return;
+			}
+			array_shift($data);
 		}
-		array_shift($data);
 		$data = implode('', $data);
 		if (method_exists($s->getCurrentTag(), 'setValue')) $s->getCurrentTag()->setValue($data);
 		else {
